Move application code out of app directory
