+++ nfl_bdb/database/models/player.py
@@ -1,0 +1,18 @@
+import datetime
+from typing import Optional
+
+from sqlalchemy.orm import Mapped, mapped_column
+
+from nfl_bdb.database.models import Base
+
+
+class Player(Base):
+    __tablename__ = "players"
+
+    player_id: Mapped[int] = mapped_column(primary_key=True)
+    name: Mapped[str] = mapped_column()
+    position: Mapped[str] = mapped_column()
+    height: Mapped[int] = mapped_column()
+    weight: Mapped[int] = mapped_column()
+    birth_date: Mapped[Optional[datetime.date]] = mapped_column(nullable=True)
+    college: Mapped[str] = mapped_column()
